requirements.txt
matplotlib==3.3.4
requests==2.23.0
numpy==1.19.5
Django==2.1
Flask==2.0.1
flask_cors==3.0.10
pyaudio==0.2.11
pydub==0.25.1
pyttsx3==2.90
baidu-aip==2.2.18.0
